LayoutTests/http/tests/security/sandboxed-iframe-javascript-top-navigation-expected.txt

CONSOLE MESSAGE: Blocked script execution in 'about:srcdoc' because the document's frame is sandboxed and the 'allow-scripts' permission is not set.
Tests that an iframe with "allow-top-navigation" but without "allow-scripts" can not navigate the top frame to a javascript URL.
